requirements.txt
PyYAML~=6.0.2
selenium~=4.27.1
webdriver-manager~=4.0.2
pytest~=8.3.3